+++ packages/css-color-parser/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix analogous components and sets for `hwb()`
 - Align precision of powerless components with the specification
+- Updated [`@csstools/css-calc`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-calc) to [`3.2.1`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-calc/CHANGELOG.md#321) (patch)
 
 ### 4.1.0